+++ tests/testthat/test-greyscale.R
@@ -7,6 +7,9 @@ test_color_class(clr_greyscale)
 test_wrong_input(clr_grayscale)
 test_wrong_input(clr_greyscale)
 
+test_zero_length_input(clr_grayscale)
+test_zero_length_input(clr_greyscale)
+
 test_that("result is grayscale", {
   methods <- c("luma", "averaging", "min_decomp",
                "max_decomp", "red_channel",