+++ tests/testthat/test-mix.R
@@ -10,6 +10,8 @@ test_severity_range(function(x, y) clr_mix(x, "blue", y))
 
 test_severity_0(function(x, y) clr_mix(x, "blue", y))
 
+test_zero_length_input(function(x) clr_mix(x, "blue"))
+
 test_that("setting shift = 1 leaves input completely changed", {
   expect_equal(clr_mix(rainbow(10), "blue", ratio = 1), color(rep("blue", 10)))
 })