+++ tests/testthat/test-rotate.R
@@ -9,3 +9,5 @@ test_severity(clr_rotate)
 test_severity_range(clr_rotate)
 
 test_severity_0(clr_rotate, tol = 1)
+
+test_zero_length_input(clr_rotate)