+++ NEWS.md
@@ -1,5 +1,7 @@
 # prismatic (development version)
 
+* All functions now accepts zero length input.
+
 # prismatic 0.2.0
 
 * Added `clr_alpha()` function to specify transparency. (#9)